deos-js-runtime: un-break the crate (6 effect_set_field usize->u64 casts) + route hex through the proven codec
Completes the DEOS panic-class fix (8be852ec8d) for the boa engine + un-breaks deos-js-runtime, which was a
committed CLASS-3 casualty: the migration changed ActionBuilder::effect_set_field to (cell, index: u64, value)
(turn/builder.rs:624) but all 6 call sites in this crate still passed usize (E0308). Cast each `slot as u64`
(applet.rs:316; world.rs:458,547,605,669,707). With the crate compiling again, runtime.rs::hex_decode_32 routes
through dregg-types parse_hex32 (the shared char-safe codec) — the boa engine's leg of the byte-slice-panic class.

Verified on persvati (native-anchor): deos-js-runtime compiles; runtime::hex_decode_32_tests
valid_hex_decodes + multibyte_64_byte_string_returns_none_not_panic both ok (2 passed, 0 failed) — the
64-byte-multibyte string returns None instead of panicking mid-codepoint in the JS callback.

+++ deos-js-runtime/src/applet.rs
@@ -313,7 +313,7 @@ impl CellApplet {
         //     world; the agent is the applet cell).
         let nonce = self.nonce();
         let action = ActionBuilder::new_unchecked_for_tests(self.cell, affordance, self.cell)
-            .effect_set_field(self.cell, slot, value)
+            .effect_set_field(self.cell, slot as u64, value)
             .effect_increment_nonce(self.cell)
             .build();
